// clients/typescript/src/api/schema.d.ts
export interface components {
    schemas: {
        BookMetadata: {
            mediaType?: components["schemas"]["MediaType"];
            /**
             * @description The title of the book
             * @example Babel
             */
            title: string;
            /**
             * @description The subtitle of the book
             * @example Or the Necessity of Violence: An Arcane History of The Oxford Translators' Revolution
             */
            subtitle?: string;
            /**
             * @deprecated
             * @description The author of the book
             * @example R. F. Kuang
             */
            author?: string;
            /**
             * @description The authors of the book
             * @example [
             *       "Susanne Valenti",
             *       "Caroline Peckham"
             *     ]
             */
            authors?: string[];
            /**
             * @deprecated
             * @description The narrator of the book
             * @example Chris Lew Kum Hoi,Billie Fulford-Brown
             */
            narrator?: string;
            /**
             * @description The narrators of the book
             * @example [
             *       "Chris Lew Kum Hoi",
             *       "Billie Fulford-Brown"
             *     ]
             */
            narrators?: string[];
            /**
             * @description The publisher of the book
             * @example HarperCollins
             */
            publisher?: string;
            /**
             * @description The year the book was published
             * @example 2022
             */
            publishedYear?: string;
            /**
             * @description The description of the book
             * @example A new dark academic fantasy by the New York Times bestselling author of The Poppy War...
             */
            description?: string;
            /**
             * @description URL to the cover image
             * @example https://example.com/books/kuang/babel/cover.jpg
             */
            cover?: string;
            /**
             * Format: isbn
             * @description The ISBN of the book
             * @example 9780063021457
             */
            isbn?: string;
            /**
             * Format: asin
             * @description The ASIN of the book
             * @example B09MD95S5V
             */
            asin?: string;
            /**
             * @description The genres of the book
             * @example [
             *       "Fantasy",
             *       "Historical Fiction",
             *       "Dark Academia"
             *     ]
             */
            genres?: string[];
            /**
             * @description The tags of the book
             * @example [
             *       "secret society",
             *       "violent",
             *       "victorian",
             *       "dark academia"
             *     ]
             */
            tags?: string[];
            series?: components["schemas"]["SeriesMetadata"][];
            /**
             * @description The language of the book
             * @example English
             */
            language?: string;
            /**
             * Format: int64
             * @description Duration in minutes
             * @example 1306
             */
            duration?: number;
        };
        SeriesMetadata: {
            /**
             * @description The series of the book
             * @example Middle Earth
             */
            series: string;
            /**
             * @description The sequence of the book in the series. Omit if there's no sequence.
             * @example 1
             */
            sequence?: string;
        };
        /**
         * @description The media type of the book
         *
         *     Clients should fall back to `unknown` for backwards compatibility
         * @enum {string}
         */
        MediaType: "book" | "audiobook" | "unknown";
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}